+++ StackExchange.Redis.Modules/RediSearch/Schema.cs
@@ -3,7 +3,7 @@
 using System;
 using System.Collections.Generic;
 
-namespace NRediSearch
+namespace StackExchange.Redis.Modules.RediSearch
 {
     /// <summary>
     /// Schema abstracts the schema definition when creating an index.
@@ -22,13 +22,11 @@ namespace NRediSearch
         {
             public String Name { get; }
             public FieldType Type { get; }
-            public bool Sortable {get;}
 
-            internal Field(string name, FieldType type, bool sortable)
+            internal Field(string name, FieldType type)
             {
                 Name = name;
                 Type = type;
-                Sortable = sortable;
             }
 
             internal virtual void SerializeRedisArgs(List<object> args)
@@ -45,17 +43,12 @@ namespace NRediSearch
                 }
                 args.Add(Name);
                 args.Add(GetForRedis(Type));
-                if(Sortable){args.Add("SORTABLE");}
             }
         }
         public class TextField : Field
         {
             public double Weight { get; }
-            internal TextField(string name, double weight = 1.0) : base(name, FieldType.FullText, false)
-            {
-                Weight = weight;
-            }
-            internal TextField(string name, bool sortable, double weight = 1.0) : base(name, FieldType.FullText, sortable)
+            internal TextField(string name, double weight = 1.0) : base(name, FieldType.FullText)
             {
                 Weight = weight;
             }
@@ -85,25 +78,13 @@ namespace NRediSearch
         }
 
         /// <summary>
-        /// Add a text field that can be sorted on
-        /// </summary>
-        /// <param name="name">the field's name</param>
-        /// <param name="weight">its weight, a positive floating point number</param>
-        /// <returns>the schema object</returns>
-        public Schema AddSortableTextField(string name, double weight = 1.0)
-        {
-            Fields.Add(new TextField(name, true, weight));
-            return this;
-        }
-
-        /// <summary>
         /// Add a numeric field to the schema
         /// </summary>
         /// <param name="name">the field's name</param>
         /// <returns>the schema object</returns>
         public Schema AddGeoField(string name)
         {
-            Fields.Add(new Field(name, FieldType.Geo, false));
+            Fields.Add(new Field(name, FieldType.Geo));
             return this;
         }
 
@@ -114,18 +95,7 @@ namespace NRediSearch
         /// <returns>the schema object</returns>
         public Schema AddNumericField(string name)
         {
-            Fields.Add(new Field(name, FieldType.Numeric, false));
-            return this;
-        }
-
-        /// <summary>
-        /// Add a numeric field that can be sorted on
-        /// </summary>
-        /// <param name="name">the field's name</param>
-        /// <returns>the schema object</returns>
-        public Schema AddSortableNumericField(string name)
-        {
-            Fields.Add(new Field(name, FieldType.Numeric, true));
+            Fields.Add(new Field(name, FieldType.Numeric));
             return this;
         }